refactor(ui, controller): 클래스 패키지 위치 변경 및 코드 위치 수정
리팩토링 내역:
  - 기존 ui 패키지에 위치했던 Validator 객체를 util 패키지로 변경
  - 컨트롤러에서 진행하던 당첨번호, 보너스 중복 여부 유효성 검사를 LottoGameSimulator 객체로 이동
